+++ src/database/SimulatedDatabaseGateway.h
@@ -23,7 +23,7 @@ public:
 
 	void insertProfile(const Profile &profile) override;
 
-	unsigned int getUserPassword(const std::string &username) override;
+	size_t getUserPassword(const std::string &username) override;
 
 private:
